stypecoup fixes
